refactor: 중복된 기능 정리
