+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/event/PaymentApprovedEvent.java
@@ -1,3 +1,3 @@
-package com.loopers.application.payment;
+package com.loopers.domain.payment.event;
 
 public record PaymentApprovedEvent(Long orderId, Long userId, Long payableAmount, Long couponId) {}
