+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/event/PaymentDeclinedEvent.java
@@ -1,3 +1,3 @@
-package com.loopers.application.payment;
+package com.loopers.domain.payment.event;
 
 public record PaymentDeclinedEvent(Long orderId) {}
